# docs/swagger.yaml
basePath: /
definitions:
  handler.ErrorResponse:
    properties:
      details:
        type: string
      error:
        type: string
    type: object
  handler.UnleashResponse:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the versioned schema of this representation of an object.
          Servers should convert recognized schemas to the latest internal value, and
          may reject unrecognized values.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources
          +optional
        type: string
      kind:
        description: |-
          Kind is a string value representing the REST resource this object represents.
          Servers may infer this from the endpoint the client submits requests to.
          Cannot be updated.
          In CamelCase.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
          +optional
        type: string
      metadata:
        $ref: '#/definitions/v1.ObjectMeta'
      spec:
        $ref: '#/definitions/unleash_nais_io_v1.UnleashSpec'
      status:
        $ref: '#/definitions/unleash_nais_io_v1.UnleashStatus'
    type: object
...
  unleash.UnleashConfig:
    properties:
      allowed-clusters:
        type: string
      allowed-namespaces:
        type: string
      allowed-teams:
        type: string
      custom-version:
        type: string
      database-pool-idle-timeout-ms:
        type: integer
      database-pool-max:
        maximum: 10
        minimum: 1
        type: integer
      enable-federation:
        type: boolean
      log-level:
        enum:
        - debug
        - info
        - warn
        - error
        - fatal
        - panic
        type: string
      name:
        type: string
    required:
    - database-pool-idle-timeout-ms
    - database-pool-max
    - log-level
    - name
    type: object
  unleash_nais_io_v1.UnleashDatabaseConfig:
    properties:
      databaseName:
        description: DatabaseName defines the name of the database to be used
        type: string
      host:
        description: Host defines the host of the database to be used
        type: string
      port:
        description: Port defines the port of the database to be used
        type: string
      secretDatabaseNameKey:
        description: SecretDatabaseNameKey is the key in the secret containing the
          database name
        type: string
      secretHostKey:
        description: SecretHostKey is the key in the secret containing the database
          host
        type: string
      secretName:
        description: SecretName is the name of the secret containing the database
          credentials
        type: string
      secretPassKey:
        description: SecretPassKey is the key in the secret containing the database
          password
        type: string
      secretPortKey:
        description: SecretPortKey is the key in the secret containing the database
          port
        type: string
      secretSSLKey:
        description: SecretSSLKey is the key in the secret containing the database
          SSL configuration
        type: string
      secretURLKey:
        description: SecretURLKey is the key in the secret containing the database
          URL
        type: string
      secretUserKey:
        description: SecretUserKey is the key in the secret containing the database
          user
        type: string
      ssl:
        description: SSL defines if the database connection should use SSL
        type: string
      url:
        description: URL defines the database URL
        type: string
      user:
        description: Username defines the username of the database to be used
        type: string
    type: object
  unleash_nais_io_v1.UnleashFederationConfig:
    properties:
      clusters:
        description: |-
          Clusters are the clusters to federate to
          +kubebuilder:validation:Optional
        items:
          type: string
        type: array
      enabled:
        description: |-
          Enable enables Unleash federation
          +kubebuilder:default=false
        type: boolean
      namespaces:
        description: |-
          Namespaces are the namespaces to federate to
          +kubebuilder:validation:Optional
        items:
          type: string
        type: array
      secretNonce:
        description: |-
          SecretNonce is the shared secret used for federation
          +kubebuilder:validation:Optional
          +kubebuilder:validation:Pattern=`^[a-z0-9]+$`
        type: string
    type: object
  unleash_nais_io_v1.UnleashIngressConfig:
    properties:
      annotations:
        additionalProperties:
          type: string
        description: |-
          Annotations is a map of annotations to add to the ingress
          +kubebuilder:validation:Optional
        type: object
      class:
        description: |-
          Class is the ingress class to use for the ingress
          +kubebuilder:validation:Optional
        type: string
      enabled:
        description: |-
          Enable enables the ingress
          +kubebuilder:default=false
        type: boolean
      host:
        description: |-
          Host is the hostname to use for the ingress
          +kubebuilder:validation:Optional
        type: string
      path:
        description: |-
          Path is the path to use for the ingress
          +kubebuilder:validation:Optional
          +kubebuilder:default="/"
        type: string
      tls:
        allOf:
        - $ref: '#/definitions/unleash_nais_io_v1.UnleashIngressTLSConfig'
        description: |-
          TLS is the TLS configuration to use for the ingress
          +kubebuilder:validation:Optional
    type: object
...
  unleash_nais_io_v1.UnleashPrometheusConfig:
    properties:
      enabled:
        description: |-
          Enable enables the prometheus metrics endpoint
          +kubebuilder:default=true
        type: boolean
    type: object
  unleash_nais_io_v1.UnleashReleaseChannelConfig:
    properties:
      name:
        description: |-
          Name is the name of the release channel
          +kubebuilder:validation:Optional
        type: string
    type: object
  unleash_nais_io_v1.UnleashSpec:
    properties:
      apiIngress:
        allOf:
        - $ref: '#/definitions/unleash_nais_io_v1.UnleashIngressConfig'
        description: |-
          ApiIngress defines the ingress for the endpoints of Unleash.
          +kubebuilder:validation:Optional
      customImage:
        description: |-
          CustomImage points to a custom image, which overrides all other version settings.
          Use at your own risk.
          +kubebuilder:validation:Optional
        type: string
      database:
        allOf:
        - $ref: '#/definitions/unleash_nais_io_v1.UnleashDatabaseConfig'
        description: |-
          Database is the database configuration.
          +kubebuilder:validation:Required
      existingServiceAccountName:
        description: |-
          ExistingServiceAccountName is the name of an already existing Kubernetes service account.
          +kubebuilder:validation:Optional
        type: string
      extraContainers:
        description: |-
          ExtraContainers is a list of extra containers to add to the deployment.
          +kubebuilder:validation:Optional
        items:
          $ref: '#/definitions/v1.Container'
        type: array
      extraEnvVars:
        description: |-
          ExtraEnvVars is a list of extra environment variables to add to the deployment.
          +kubebuilder:validation:Optional
        items:
          $ref: '#/definitions/v1.EnvVar'
        type: array
      extraVolumeMounts:
        description: |-
          ExtraVolumeMounts is a list of extra volume mounts to add to the deployment.
          +kubebuilder:validation:Optional
        items:
          $ref: '#/definitions/v1.VolumeMount'
        type: array
      extraVolumes:
        description: |-
          ExtraVolumes is a list of extra volume mounts to add to the deployment.
          +kubebuilder:validation:Optional
        items:
          $ref: '#/definitions/v1.Volume'
        type: array
      federation:
        allOf:
        - $ref: '#/definitions/unleash_nais_io_v1.UnleashFederationConfig'
        description: |-
          Federation is the configuration for Unleash federation.
          +kubebuilder:validation:Optional
      networkPolicy:
        allOf:
        - $ref: '#/definitions/unleash_nais_io_v1.UnleashNetworkPolicyConfig'
        description: |-
          NetworkPolicy defines the network policy configuration.
          +kubebuilder:validation:Optional
      podAnnotations:
        additionalProperties:
          type: string
        description: |-
          PodAnnotations are additional annotations to add to the Unleash pods.
          +kubebuilder:validation:Optional
        type: object
      podLabels:
        additionalProperties:
          type: string
        description: |-
          PodLabels are additional labels to add to the Unleash pods.
          +kubebuilder:validation:Optional
        type: object
      prometheus:
        allOf:
        - $ref: '#/definitions/unleash_nais_io_v1.UnleashPrometheusConfig'
        description: |-
          Prometheus defines the Prometheus metrics collection configuration.
          +kubebuilder:validation:Optional
      releaseChannel:
        allOf:
        - $ref: '#/definitions/unleash_nais_io_v1.UnleashReleaseChannelConfig'
        description: |-
          ReleaseChannel is the release channel to use for the Unleash deployment.
          +kubebuilder:validation:Optional
      resources:
        allOf:
        - $ref: '#/definitions/v1.ResourceRequirements'
        description: |-
          Resources are the resource requests and limits for the Unleash deployment.
          +kubebuilder:validation:Optional
          +kubebuilder:default:={requests: {cpu: "300m", memory: "256Mi"}, limits: { memory: "512Mi"}}
      size:
        description: |-
          Size is the size of the Unleash deployment.
          +kubebuilder:default=1
        type: integer
      webIngress:
        allOf:
        - $ref: '#/definitions/unleash_nais_io_v1.UnleashIngressConfig'
        description: |-
          WebIngress defines the ingress configuration for the web interface.
          +kubebuilder:validation:Optional
    type: object
  unleash_nais_io_v1.UnleashStatus:
    properties:
      conditions:
        description: |-
          Represents the observations of a Unleash's current state.
          Unleash.status.conditions.type are: "Reconciled", "Connected", and "Degraded"
          Unleash.status.conditions.status are one of True, False, Unknown.
          Unleash.status.conditions.reason the value should be a CamelCase string and producers of specific
          condition types may define expected values and meanings for this field, and whether the values
          are considered a guaranteed API.
          Unleash.status.conditions.Message is a human readable message indicating details about the transition.
        items:
          $ref: '#/definitions/v1.Condition'
        type: array
      connected:
        description: |-
          Connected is true when the Unleash resource has been connected to the Unleash server
          This is used for kubectl printing purposes. Rather than relying on this
          value, check the conditions instead.
          +kubebuilder:default=false
        type: boolean
      reconciled:
        description: |-
          Reconciled is true when the Unleash resources hav been reconciled
          successfully.
          This is used for kubectl printing purposes. Rather than relying on this
          value, check the conditions instead.
          +kubebuilder:default=false
        type: boolean
      releaseChannelName:
        description: |-
          ReleaseChannelName tracks which ReleaseChannel was used to resolve the image
          This ensures we only use the resolved image if it matches the current ReleaseChannel
          +kubebuilder:validation:Optional
        type: string
      resolvedReleaseChannelImage:
        description: |-
          ResolvedReleaseChannelImage is the image resolved from the ReleaseChannel
          This is set by the controller when a ReleaseChannel is specified and tracks
          the current image that should be used for this Unleash instance.
          +kubebuilder:validation:Optional
        type: string
      version:
        description: |-
          Version is the reported version of the Unleash server
          +kubebuilder:default="unknown"
        type: string
    type: object
...
  v1.EnvVar:
    properties:
      name:
        description: |-
          Name of the environment variable.
          May consist of any printable ASCII characters except '='.
        type: string
      value:
        description: |-
          Variable references $(VAR_NAME) are expanded
          using the previously defined environment variables in the container and
          any service environment variables. If a variable cannot be resolved,
          the reference in the input string will be unchanged. Double $$ are reduced
          to a single $, which allows for escaping the $(VAR_NAME) syntax: i.e.
          "$$(VAR_NAME)" will produce the string literal "$(VAR_NAME)".
          Escaped references will never be expanded, regardless of whether the variable
          exists or not.
          Defaults to "".
          +optional
        type: string
      valueFrom:
        allOf:
        - $ref: '#/definitions/v1.EnvVarSource'
        description: |-
          Source for the environment variable's value. Cannot be used if value is not empty.
          +optional
    type: object
...
  v1.FieldsV1:
    type: object
...
  v1.ManagedFieldsEntry:
    properties:
      apiVersion:
        description: |-
          APIVersion defines the version of this resource that this field set
          applies to. The format is "group/version" just like the top-level
          APIVersion field. It is necessary to track the version of a field
          set because it cannot be automatically converted.
        type: string
      fieldsType:
        description: |-
          FieldsType is the discriminator for the different fields format and version.
          There is currently only one possible value: "FieldsV1"
        type: string
      fieldsV1:
        allOf:
        - $ref: '#/definitions/v1.FieldsV1'
        description: |-
          FieldsV1 holds the first JSON version format as described in the "FieldsV1" type.
          +optional
      manager:
        description: Manager is an identifier of the workflow managing these fields.
        type: string
      operation:
        allOf:
        - $ref: '#/definitions/v1.ManagedFieldsOperationType'
        description: |-
          Operation is the type of operation which lead to this ManagedFieldsEntry being created.
          The only valid values for this field are 'Apply' and 'Update'.
      subresource:
        description: |-
          Subresource is the name of the subresource used to update that object, or
          empty string if the object was updated through the main resource. The
          value of this field is used to distinguish between managers, even if they
          share the same name. For example, a status update will be distinct from a
          regular update using the same manager name.
          Note that the APIVersion field is not related to the Subresource field and
          it always corresponds to the version of the main resource.
        type: string
      time:
        description: |-
          Time is the timestamp of when the ManagedFields entry was added. The
          timestamp will also be updated if a field is added, the manager
          changes any of the owned fields value or removes a field. The
          timestamp does not update when a field is removed from the entry
          because another manager took it over.
          +optional
        type: string
    type: object
  v1.ManagedFieldsOperationType:
    enum:
    - Apply
    - Update
    type: string
    x-enum-varnames:
    - ManagedFieldsOperationApply
    - ManagedFieldsOperationUpdate
...
  v1.ObjectMeta:
    properties:
      annotations:
        additionalProperties:
          type: string
        description: |-
          Annotations is an unstructured key value map stored with a resource that may be
          set by external tools to store and retrieve arbitrary metadata. They are not
          queryable and should be preserved when modifying objects.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/annotations
          +optional
        type: object
      creationTimestamp:
        description: |-
          CreationTimestamp is a timestamp representing the server time when this object was
          created. It is not guaranteed to be set in happens-before order across separate operations.
          Clients may not set this value. It is represented in RFC3339 form and is in UTC.

          Populated by the system.
          Read-only.
          Null for lists.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      deletionGracePeriodSeconds:
        description: |-
          Number of seconds allowed for this object to gracefully terminate before
          it will be removed from the system. Only set when deletionTimestamp is also set.
          May only be shortened.
          Read-only.
          +optional
        type: integer
      deletionTimestamp:
        description: |-
          DeletionTimestamp is RFC 3339 date and time at which this resource will be deleted. This
          field is set by the server when a graceful deletion is requested by the user, and is not
          directly settable by a client. The resource is expected to be deleted (no longer visible
          from resource lists, and not reachable by name) after the time in this field, once the
          finalizers list is empty. As long as the finalizers list contains items, deletion is blocked.
          Once the deletionTimestamp is set, this value may not be unset or be set further into the
          future, although it may be shortened or the resource may be deleted prior to this time.
          For example, a user may request that a pod is deleted in 30 seconds. The Kubelet will react
          by sending a graceful termination signal to the containers in the pod. After that 30 seconds,
          the Kubelet will send a hard termination signal (SIGKILL) to the container and after cleanup,
          remove the pod from the API. In the presence of network partitions, this object may still
          exist after this timestamp, until an administrator or automated process can determine the
          resource is fully terminated.
          If not set, graceful deletion of the object has not been requested.

          Populated by the system when a graceful deletion is requested.
          Read-only.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#metadata
          +optional
        type: string
      finalizers:
        description: |-
          Must be empty before the object is deleted from the registry. Each entry
          is an identifier for the responsible component that will remove the entry
          from the list. If the deletionTimestamp of the object is non-nil, entries
          in this list can only be removed.
          Finalizers may be processed and removed in any order.  Order is NOT enforced
          because it introduces significant risk of stuck finalizers.
          finalizers is a shared field, any actor with permission can reorder it.
          If the finalizer list is processed in order, then this can lead to a situation
          in which the component responsible for the first finalizer in the list is
          waiting for a signal (field value, external system, or other) produced by a
          component responsible for a finalizer later in the list, resulting in a deadlock.
          Without enforced ordering finalizers are free to order amongst themselves and
          are not vulnerable to ordering changes in the list.
          +optional
          +patchStrategy=merge
          +listType=set
        items:
          type: string
        type: array
      generateName:
        description: |-
          GenerateName is an optional prefix, used by the server, to generate a unique
          name ONLY IF the Name field has not been provided.
          If this field is used, the name returned to the client will be different
          than the name passed. This value will also be combined with a unique suffix.
          The provided value has the same validation rules as the Name field,
          and may be truncated by the length of the suffix required to make the value
          unique on the server.

          If this field is specified and the generated name exists, the server will return a 409.

          Applied only if Name is not specified.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#idempotency
          +optional
        type: string
      generation:
        description: |-
          A sequence number representing a specific generation of the desired state.
          Populated by the system. Read-only.
          +optional
        type: integer
      labels:
        additionalProperties:
          type: string
        description: |-
          Map of string keys and values that can be used to organize and categorize
          (scope and select) objects. May match selectors of replication controllers
          and services.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/labels
          +optional
        type: object
      managedFields:
        description: |-
          ManagedFields maps workflow-id and version to the set of fields
          that are managed by that workflow. This is mostly for internal
          housekeeping, and users typically shouldn't need to set or
          understand this field. A workflow can be the user's name, a
          controller's name, or the name of a specific apply path like
          "ci-cd". The set of fields is always in the version that the
          workflow used when modifying the object.

          +optional
          +listType=atomic
        items:
          $ref: '#/definitions/v1.ManagedFieldsEntry'
        type: array
      name:
        description: |-
          Name must be unique within a namespace. Is required when creating resources, although
          some resources may allow a client to request the generation of an appropriate name
          automatically. Name is primarily intended for creation idempotence and configuration
          definition.
          Cannot be updated.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#names
          +optional
        type: string
      namespace:
        description: |-
          Namespace defines the space within which each name must be unique. An empty namespace is
          equivalent to the "default" namespace, but "default" is the canonical representation.
          Not all objects are required to be scoped to a namespace - the value of this field for
          those objects will be empty.

          Must be a DNS_LABEL.
          Cannot be updated.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/namespaces
          +optional
        type: string
      ownerReferences:
        description: |-
          List of objects depended by this object. If ALL objects in the list have
          been deleted, this object will be garbage collected. If this object is managed by a controller,
          then an entry in this list will point to this controller, with the controller field set to true.
          There cannot be more than one managing controller.
          +optional
          +patchMergeKey=uid
          +patchStrategy=merge
          +listType=map
          +listMapKey=uid
        items:
          $ref: '#/definitions/v1.OwnerReference'
        type: array
      resourceVersion:
        description: |-
          An opaque value that represents the internal version of this object that can
          be used by clients to determine when objects have changed. May be used for optimistic
          concurrency, change detection, and the watch operation on a resource or set of resources.
          Clients must treat these values as opaque and passed unmodified back to the server.
          They may only be valid for a particular resource or set of resources.

          Populated by the system.
          Read-only.
          Value must be treated as opaque by clients and .
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency
          +optional
        type: string
      selfLink:
        description: |-
          Deprecated: selfLink is a legacy read-only field that is no longer populated by the system.
          +optional
        type: string
      uid:
        description: |-
          UID is the unique in time and space value for this object. It is typically generated by
          the server on successful creation of a resource and is not allowed to change on PUT
          operations.

          Populated by the system.
          Read-only.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#uids
          +optional
        type: string
    type: object
  v1.OwnerReference:
    properties:
      apiVersion:
        description: API version of the referent.
        type: string
      blockOwnerDeletion:
        description: |-
          If true, AND if the owner has the "foregroundDeletion" finalizer, then
          the owner cannot be deleted from the key-value store until this
          reference is removed.
          See https://kubernetes.io/docs/concepts/architecture/garbage-collection/#foreground-deletion
          for how the garbage collector interacts with this field and enforces the foreground deletion.
          Defaults to false.
          To set this field, a user needs "delete" permission of the owner,
          otherwise 422 (Unprocessable Entity) will be returned.
          +optional
        type: boolean
      controller:
        description: |-
          If true, this reference points to the managing controller.
          +optional
        type: boolean
      kind:
        description: |-
          Kind of the referent.
          More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds
        type: string
      name:
        description: |-
          Name of the referent.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#names
        type: string
      uid:
        description: |-
          UID of the referent.
          More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#uids
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: NAIS Team
    url: https://github.com/nais/bifrost
  description: |-
    API for managing Unleash (feature flag) instances
    This service provides a JSON API for creating, updating, and deleting Unleash instances.
  license:
    name: MIT
    url: https://github.com/nais/bifrost/blob/main/LICENSE
  title: Bifrost API
  version: "1.0"
paths:
...
  /unleash/{id}:
    delete:
      description: Delete an existing Unleash instance by name
      parameters:
      - description: Instance name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted
        "404":
          description: Instance not found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Delete Unleash instance
      tags:
      - unleash
  /unleash/{id}/edit:
    post:
      consumes:
      - application/json
      description: Create a new Unleash instance or update an existing one
      parameters:
      - description: Unleash configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/unleash.UnleashConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created or updated
          schema:
            $ref: '#/definitions/handler.UnleashResponse'
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create or update Unleash instance
      tags:
      - unleash
  /unleash/new:
    post:
      consumes:
      - application/json
      description: Create a new Unleash instance or update an existing one
      parameters:
      - description: Unleash configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/unleash.UnleashConfig'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created or updated
          schema:
            $ref: '#/definitions/handler.UnleashResponse'
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Create or update Unleash instance
      tags:
      - unleash
swagger: "2.0"